Updated the return type of confidenceScore in Dialogue; Implemented StepSuggestionInteractor for our second use case partially. Still need to implement the part for accessing parsed transcript and outputing the suggestions.

+++ backend/src/main/java/com/speakfluid/backend/entities/Dialogue.java
@@ -33,7 +33,7 @@ public class Dialogue {
     public ArrayList<Speech> getUserMessage(){
         return this.userMessage;
     }
-    public void setConfidenceScore(int confidenceScore) {
+    public void setConfidenceScore(double confidenceScore) {
         this.confidenceScore = confidenceScore;
     }
